Replace hard-coded secrets by randomly-generated values.
