+++ pkg/util/util.go
@@ -1,0 +1,16 @@
+package util
+
+import (
+	"crypto/rand"
+	"log"
+)
+
+// SecureRandomBytes returns the requested number of bytes using crypto/rand
+func SecureRandomBytes(length int) []byte {
+	var randomBytes = make([]byte, length)
+	_, err := rand.Read(randomBytes)
+	if err != nil {
+		log.Fatal("Unable to generate random bytes")
+	}
+	return randomBytes
+}
